mapper, support: clean up a bit
